Move UserAvatar to reviewing/components
