+++ indico/modules/events/client/js/reviewing/components/UserAvatar.jsx
